refactor: separate into serenity handler and call commands struct Commands
